# docs/trading_alerts_openapi_compress.yaml
openapi: 3.0.3
info:
  title: Trading Alerts SaaS API
  version: 7.1.0
  description: |
    Commercial SaaS platform for trading alerts and signal generation with real-time chart visualization.
    **Stack:** Next.js 15 + Flask MT5 + Dual Payment (Stripe/dLocal)
    **Auth:** NextAuth.js v4 (Google OAuth + Email/Password, JWT sessions)
    **Tiers:** FREE (5 symbols × 3 timeframes = 15 combos), PRO (15 symbols × 9 timeframes = 135 combos)
    **Payment:** Stripe (international, auto-renewal, 7-day trial), dLocal (8 markets: IN, NG, PK, VN, ID, TH, ZA, TR, manual renewal)
  contact:
    name: API Support
    email: support@tradingalerts.com
  license:
    name: Proprietary
    url: https://tradingalerts.com/license
servers:
  - url: https://api.tradingalerts.com
    description: Production server
  - url: https://staging-api.tradingalerts.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates new user account with FREE tier by default
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/alerts:
    get:
      tags:
        - Alerts
      summary: Get user's alerts
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: symbol
          in: query
          schema:
            type: string
          description: Filter by symbol
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'

    post:
      tags:
        - Alerts
      summary: Create a new alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Alert limit reached for tier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - User
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                image:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Get all users (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
        - name: tier
          in: query
          schema:
            $ref: '#/components/schemas/UserTier'
          description: Filter by tier
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of users with tier distribution
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserAdmin'
                  total:
                    type: integer
                  freeCount:
                    type: integer
                  proCount:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login (includes tier information)
...
  schemas:
...
    UserTier:
      type: string
      enum:
        - FREE
        - PRO
      description: |
        FREE: 5 symbols × 3 timeframes = 15 combinations
        PRO: 15 symbols × 9 timeframes = 135 combinations
    Symbol:
      type: string
      enum:
        - AUDJPY
        - AUDUSD
        - BTCUSD
        - ETHUSD
        - EURUSD
        - GBPJPY
        - GBPUSD
        - NDX100
        - NZDUSD
        - US30
        - USDCAD
        - USDCHF
        - USDJPY
        - XAGUSD
        - XAUUSD
      description: |
        15 total symbols. FREE: BTCUSD, EURUSD, USDJPY, US30, XAUUSD. PRO: All 15.
    Timeframe:
      type: string
      enum:
        - M5
        - M15
        - M30
        - H1
        - H2
        - H4
        - H8
        - H12
        - D1
      description: |
        9 total timeframes. FREE: H1, H4, D1. PRO: All 9 (M5 and H12 are PRO-only).
    AlertType:
      type: string
      enum:
        - PRICE_TOUCH_LINE
        - PRICE_CROSS_LINE
        - FRACTAL_NEW
        - LINE_BREAK
        - INDICATOR_VALUE
        - CUSTOM
...
    SubscriptionStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - CANCELED
        - PAST_DUE
        - UNPAID
        - TRIALING

    # ============================================
    # AUTHENTICATION SCHEMAS
    # ============================================
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: ^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).*$
          description: Must contain uppercase, lowercase, and number
        acceptTerms:
          type: boolean
          enum: [true]
      required:
        - name
        - email
        - password
        - acceptTerms
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
            tier:
              $ref: '#/components/schemas/UserTier'
...
    Alert:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
          nullable: true
        symbol:
          $ref: '#/components/schemas/Symbol'
        timeframe:
          $ref: '#/components/schemas/Timeframe'
        alertType:
          $ref: '#/components/schemas/AlertType'
        condition:
          type: object
        isActive:
          type: boolean
        lastTriggered:
          type: string
          format: date-time
          nullable: true
        triggerCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAlertRequest:
      type: object
      properties:
        name:
          type: string
        symbol:
          $ref: '#/components/schemas/Symbol'
        timeframe:
          $ref: '#/components/schemas/Timeframe'
        alertType:
          $ref: '#/components/schemas/AlertType'
        condition:
          type: object
      required:
        - symbol
        - timeframe
        - alertType
        - condition
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
          description: Profile picture URL
        tier:
          $ref: '#/components/schemas/UserTier'
        emailVerified:
          type: string
          format: date-time
          nullable: true
          description: |
            Verification timestamp. Email/password users: set after verification. Google OAuth: auto-set. Critical for security.
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
        authMethod:
          type: string
          enum: [credentials, google, both]
          description: credentials=email/password only, google=OAuth only, both=linked
...
    UserAdmin:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            passwordHash:
              type: string
              nullable: true
              description: Password hash for email/password auth. NULL for OAuth-only users.
            subscription:
              type: object
              nullable: true
              properties:
                status:
                  $ref: '#/components/schemas/SubscriptionStatus'
                plan:
                  $ref: '#/components/schemas/UserTier'
            accounts:
              type: array
              items:
                $ref: '#/components/schemas/Account'
    Account:
      type: object
      description: OAuth provider account linked to user
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [oauth]
        provider:
          type: string
          enum: [google]
        providerAccountId:
          type: string
        refresh_token:
          type: string
          nullable: true
        access_token:
          type: string
          nullable: true
        expires_at:
          type: integer
          nullable: true
        token_type:
          type: string
          nullable: true
        scope:
          type: string
          nullable: true
        id_token:
          type: string
          nullable: true
        session_state:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - userId
        - type
        - provider
        - providerAccountId

    # ============================================
    # ADMIN SCHEMAS
    # ============================================
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true
        details:
          type: object
          nullable: true
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
...